# api/swagger.yaml
basePath: /beetle
definitions:
  cloudmodel.CloudProperty:
    properties:
      csp:
        example: aws
        type: string
      region:
        example: ap-northeast-2
        type: string
    type: object
...
  controller.MigrateObjectStorageRequest:
    properties:
      description:
        type: string
      status:
        type: string
      targetCloud:
        $ref: '#/definitions/cloudmodel.CloudProperty'
      targetObjectStorages:
        items:
          $ref: '#/definitions/controller.TargetObjectStorageProperty'
        type: array
    type: object
...
  controller.RecommendObjectStorageRequest:
    properties:
      desiredCloud:
        $ref: '#/definitions/cloudmodel.CloudProperty'
      sourceObjectStorages:
        items:
          $ref: '#/definitions/controller.SourceObjectStorageProperty'
        minItems: 1
        type: array
    required:
    - desiredCloud
    - sourceObjectStorages
    type: object
...
  controller.SourceObjectStorageProperty:
    properties:
      accessFrequency:
        description: Access pattern (critical for storage class selection)
        type: string
      bucketName:
        description: Basic identification
        type: string
      corsEnabled:
        description: Whether CORS is enabled
        type: boolean
      corsRules:
        description: CORS rules configuration
        items:
          $ref: '#/definitions/controller.CORSRule'
        type: array
      creationDate:
        description: Creation date (RFC3339)
        type: string
      encryptionEnabled:
        description: Security settings
        type: boolean
      isPublic:
        description: Whether bucket has public access
        type: boolean
      objectCount:
        description: Total number of objects
        type: integer
      tags:
        additionalProperties:
          type: string
        description: Metadata
        type: object
      totalSizeBytes:
        description: Capacity information (for cost estimation and recommendations)
        type: integer
      versioningEnabled:
        description: Feature settings
        type: boolean
    required:
    - bucketName
    type: object
  controller.TargetObjectStorageProperty:
    properties:
      bucketName:
        description: Recommended target bucket name with deterministic suffix
        type: string
      corsEnabled:
        description: Whether CORS is configured
        type: boolean
      corsRules:
        description: CORS rules configuration
        items:
          $ref: '#/definitions/controller.CORSRule'
        type: array
      sourceBucketName:
        description: Source bucket name for referencing
        type: string
      versioningEnabled:
        description: Whether to enable versioning
        type: boolean
    type: object
...
  model.ApiResponse-any:
    properties:
      data:
        description: Contains the actual response data (single object, list, or page)
      error:
        description: Error message for failed responses
        example: Error message if failure
        type: string
      message:
        description: Optional message for additional context
        example: Operation successful
        type: string
      success:
        description: Indicates whether the API call was successful
        example: true
        type: boolean
    type: object
...
  model.ApiResponse-controller_ObjectStorageInfo:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/controller.ObjectStorageInfo'
        description: Contains the actual response data (single object, list, or page)
      error:
        description: Error message for failed responses
        example: Error message if failure
        type: string
      message:
        description: Optional message for additional context
        example: Operation successful
        type: string
      success:
        description: Indicates whether the API call was successful
        example: true
        type: boolean
    type: object
...
  model.ApiResponse-tbclient_ListAllMyBucketsResult:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/tbclient.ListAllMyBucketsResult'
        description: Contains the actual response data (single object, list, or page)
      error:
        description: Error message for failed responses
        example: Error message if failure
        type: string
      message:
        description: Optional message for additional context
        example: Operation successful
        type: string
      success:
        description: Indicates whether the API call was successful
        example: true
        type: boolean
    type: object
  model.ApiResponse-tbclient_ListBucketResult:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/tbclient.ListBucketResult'
        description: Contains the actual response data (single object, list, or page)
      error:
        description: Error message for failed responses
        example: Error message if failure
        type: string
      message:
        description: Optional message for additional context
        example: Operation successful
        type: string
      success:
        description: Indicates whether the API call was successful
        example: true
        type: boolean
    type: object
...
  tbclient.Buckets:
    properties:
      bucket:
        items:
          $ref: '#/definitions/tbclient.Bucket'
        type: array
    type: object
  tbclient.ListAllMyBucketsResult:
    properties:
      buckets:
        $ref: '#/definitions/tbclient.Buckets'
      owner:
        $ref: '#/definitions/tbclient.Owner'
    type: object
  tbclient.ListBucketResult:
    properties:
      isTruncated:
        example: false
        type: boolean
      marker:
        example: ""
        type: string
      maxKeys:
        example: 1000
        type: integer
      name:
        example: spider-test-bucket
        type: string
      prefix:
        example: ""
        type: string
    type: object
...
host: localhost:8056
info:
  contact:
    name: API Support
    url: https://github.com/cloud-barista/cm-beetle/issues/new/choose
  description: CM-Beetle recommends optimal target cloud infrastructure and executes
    migration.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: CM-Beetle API
  version: latest
paths:
...
  /migration/middleware/objectStorage:
    get:
      consumes:
      - application/json
      description: |-
        Get the list of all object storages (buckets) in the specified cloud service provider and region

        [Note] Connection name format: `{csp}-{region}` (e.g., aws-ap-northeast-2)
      operationId: ListObjectStorages
      parameters:
      - default: aws
        description: Cloud service provider
        enum:
        - aws
        - alibaba
        in: query
        name: csp
        required: true
        type: string
      - default: ap-northeast-2
        description: Cloud region
        in: query
        name: region
        required: true
        type: string
      - description: Unique request ID (auto-generated if not provided). Used for
          tracking request status and correlating logs.
        in: header
        name: X-Request-Id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved object storage list
          schema:
            $ref: '#/definitions/model.ApiResponse-tbclient_ListAllMyBucketsResult'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/model.ApiResponse-any'
        "500":
          description: Internal server error during list operation
          schema:
            $ref: '#/definitions/model.ApiResponse-any'
      summary: List object storages (buckets)
      tags:
      - '[Migration] Managed middleware (preview)'
    post:
      consumes:
      - application/json
      description: |
        Migrate object storages to cloud based on recommendation results

        [Note] This API creates object storages (buckets) in the target cloud.
        - Input should be the output from RecommendObjectStorage API
        - Connection name format: `{csp}-{region}` (e.g., aws-ap-northeast-2)

        [Note]
        * Examples(test result): https://github.com/cloud-barista/cm-beetle/blob/main/docs/test-results-data-migration.md
      operationId: MigrateObjectStorage
      parameters:
      - description: Object storage migration request (use RecommendObjectStorage
          response)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.MigrateObjectStorageRequest'
      - description: Unique request ID (auto-generated if not provided). Used for
          tracking request status and correlating logs.
        in: header
        name: X-Request-Id
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created - Object storages created successfully
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/model.ApiResponse-any'
        "500":
          description: Internal server error during object storage creation
          schema:
            $ref: '#/definitions/model.ApiResponse-any'
      summary: Migrate object storages to cloud
      tags:
      - '[Migration] Managed middleware (preview)'
  /migration/middleware/objectStorage/{objectStorageName}:
    delete:
      consumes:
      - application/json
      description: |-
        Delete a specific object storage (bucket)

        [Note]
        - Connection name format: `{csp}-{region}` (e.g., aws-ap-northeast-2)
        - The bucket must be empty before deletion
      operationId: DeleteObjectStorage
      parameters:
      - description: Object Storage Name (bucket name)
        in: path
        name: objectStorageName
        required: true
        type: string
      - default: aws
        description: Cloud service provider
        enum:
        - aws
        - alibaba
        in: query
        name: csp
        required: true
        type: string
      - default: ap-northeast-2
        description: Cloud region
        in: query
        name: region
        required: true
        type: string
      - description: Unique request ID (auto-generated if not provided). Used for
          tracking request status and correlating logs.
        in: header
        name: X-Request-Id
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Object storage deleted successfully
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/model.ApiResponse-any'
        "404":
          description: Object storage not found
          schema:
            $ref: '#/definitions/model.ApiResponse-any'
        "500":
          description: Internal server error during deletion
          schema:
            $ref: '#/definitions/model.ApiResponse-any'
      summary: Delete object storage (bucket)
      tags:
      - '[Migration] Managed middleware (preview)'
    get:
      consumes:
      - application/json
      description: |-
        Get details of a specific object storage (bucket)

        [Note] Connection name format: `{csp}-{region}` (e.g., aws-ap-northeast-2)
      operationId: GetObjectStorage
      parameters:
      - description: Object Storage Name (bucket name)
        in: path
        name: objectStorageName
        required: true
        type: string
      - default: aws
        description: Cloud service provider
        enum:
        - aws
        - alibaba
        in: query
        name: csp
        required: true
        type: string
      - default: ap-northeast-2
        description: Cloud region
        in: query
        name: region
        required: true
        type: string
      - description: Unique request ID (auto-generated if not provided). Used for
          tracking request status and correlating logs.
        in: header
        name: X-Request-Id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved object storage details
          schema:
            $ref: '#/definitions/model.ApiResponse-tbclient_ListBucketResult'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/model.ApiResponse-any'
        "404":
          description: Object storage not found
          schema:
            $ref: '#/definitions/model.ApiResponse-any'
        "500":
          description: Internal server error during get operation
          schema:
            $ref: '#/definitions/model.ApiResponse-any'
      summary: Get object storage (bucket) details
      tags:
      - '[Migration] Managed middleware (preview)'
    head:
      consumes:
      - application/json
      description: |-
        Check if a specific object storage (bucket) exists

        [Note]
        - Connection name format: `{csp}-{region}` (e.g., aws-ap-northeast-2)
        - Returns 200 OK if the bucket exists, 404 Not Found if it doesn't exist
      operationId: ExistObjectStorage
      parameters:
      - description: Object Storage Name (bucket name)
        in: path
        name: objectStorageName
        required: true
        type: string
      - default: aws
        description: Cloud service provider
        enum:
        - aws
        - alibaba
        in: query
        name: csp
        required: true
        type: string
      - default: ap-northeast-2
        description: Cloud region
        in: query
        name: region
        required: true
        type: string
      - description: Unique request ID (auto-generated if not provided). Used for
          tracking request status and correlating logs.
        in: header
        name: X-Request-Id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK - Object storage exists
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/model.ApiResponse-any'
        "404":
          description: Object storage not found
          schema:
            $ref: '#/definitions/model.ApiResponse-any'
        "500":
          description: Internal server error during existence check
          schema:
            $ref: '#/definitions/model.ApiResponse-any'
      summary: Check object storage (bucket) existence
      tags:
      - '[Migration] Managed middleware (preview)'
...
  /recommendation/middleware/objectStorage:
    post:
      consumes:
      - application/json
      description: |-
        Recommend an appropriate object storage for cloud migration

        [Note] `desiredCsp` and `desiredRegion` are required.
        - `desiredCsp` and `desiredRegion` can set on the query parameter or the request body.

        - If desiredCsp and desiredRegion are set on request body, the values in the query parameter will be ignored.

        [Warning] the recommended bucket name may be globally unique.
        - Beetle supports adding a suffix based on the existing bucket name to ensure uniqueness.
        - Suppose that the existing bucket name is unique enough.
        - Generate a suffix based on the existing bucket name.
        - e.g., "my-bucket" -> SHA256 hash -> base64 URL-safe encoding (6 bytes) -> lowercase -> "my-bucket-{suffix}"
      operationId: RecommendObjectStorage
      parameters:
      - description: Specify the your object storage to be migrated
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.RecommendObjectStorageRequest'
      - default: aws
        description: CSP (e.g., aws, azure, gcp)
        enum:
        - aws
        - azure
        - gcp
        - alibaba
        - ncp
        in: query
        name: desiredCsp
        type: string
      - default: ap-northeast-2
        description: Region (e.g., ap-northeast-2)
        in: query
        name: desiredRegion
        type: string
      - description: Unique request ID (auto-generated if not provided). Used for
          tracking request status and correlating logs.
        in: header
        name: X-Request-Id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully recommended object storage
          schema:
            $ref: '#/definitions/model.ApiResponse-controller_ObjectStorageInfo'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/model.ApiResponse-any'
        "500":
          description: Internal server error during recommendation
          schema:
            $ref: '#/definitions/model.ApiResponse-any'
      summary: Recommend an object storage for cloud migration
      tags:
      - '[Recommendation] Managed middleware (preview)'
...
security:
- BasicAuth: []
- BearerAuth: []
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    description: Enter your bearer token in the format **Bearer &lt;token&gt;**
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...